requirements.txt
streamlit
boto3
opencv-python-headless
pandas
matplotlib
Pillow
